Extract AthleticsEventCard in single file [#3867]
